+++ argopy/stores/implementations/http.py
@@ -516,14 +516,14 @@ class httpstore(ArgoStoreProto):
                         # log.info('drop_variables_not_in_all_datasets')
                         ds_list = drop_variables_not_in_all_datasets(ds_list)
 
-                    log.info("Nb of dataset to concat: %i" % len(ds_list))
+                    # log.info("Nb of dataset to concat: %i" % len(ds_list))
                     # log.debug(concat_dim)
                     # for ds in ds_list:
                     #     log.debug(ds[concat_dim])
-                    log.info(
-                        "Dataset sizes before concat: %s"
-                        % [len(ds[concat_dim]) for ds in ds_list]
-                    )
+                    # log.info(
+                    #     "Dataset sizes before concat: %s"
+                    #     % [len(ds[concat_dim]) for ds in ds_list]
+                    # )
                     ds = xr.concat(
                         ds_list,
                         dim=concat_dim,
@@ -531,7 +531,7 @@ class httpstore(ArgoStoreProto):
                         coords="all",
                         compat="override",    # skip comparing and pick variable from first dataset
                     )
-                    log.info("Dataset size after concat: %i" % len(ds[concat_dim]))
+                    # log.info("Dataset size after concat: %i" % len(ds[concat_dim]))
                     return ds, True
                 else:
                     ds_list = [v for v in dict(sorted(obj_list.items())).values()]