+++ docs/whats-new.rst
@@ -14,7 +14,7 @@ Coming up next (un-released)
 Features and front-end API
 ^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-- **New** :class:`ArgoSensor` class to work with Argo sensor models, types and floats equipped with. Check the new :ref:`Argo sensor: models and types` documentation section. (:pr:`532`) by |gmaze|.
+- **New** :class:`ArgoSensor` class to work with Argo sensor models, types and floats equipped with. Check the new :ref:`Argo sensors` documentation section. (:pr:`532`) by |gmaze|.
 
 
 v1.3.1 (22 Oct. 2025)